+++ packages/typespec-autorest-canonical/test/shared-routes.test.ts
@@ -1,7 +1,7 @@
+import { OpenAPI2Document } from "@azure-tools/typespec-autorest";
 import { HttpVerb } from "@typespec/http";
 import { deepStrictEqual, ok, strictEqual } from "assert";
 import { it } from "vitest";
-import { OpenAPI2Document } from "../src/types.js";
 import { openApiFor } from "./test-host.js";
 
 interface ExpectOperation {
