+++ packages/typespec-autorest/src/openapi2-document.ts
