+++ packages/typespec-autorest/test/parameters.test.ts
@@ -1,7 +1,7 @@
 import { expectDiagnostics } from "@typespec/compiler/testing";
 import { deepStrictEqual, ok, strictEqual } from "assert";
 import { describe, expect, it } from "vitest";
-import { OpenAPI2HeaderParameter, OpenAPI2QueryParameter } from "../src/types.js";
+import { OpenAPI2HeaderParameter, OpenAPI2QueryParameter } from "../src/openapi2-document.js";
 import { diagnoseOpenApiFor, ignoreUseStandardOps, openApiFor } from "./test-host.js";
 
 describe("typespec-autorest: parameters", () => {
