+++ packages/typespec-autorest/test/primitive-types.test.ts
@@ -1,7 +1,7 @@
 import { expectDiagnostics } from "@typespec/compiler/testing";
 import { deepStrictEqual, ok } from "assert";
 import { describe, it } from "vitest";
-import { OpenAPI2Parameter, OpenAPI2Schema } from "../src/types.js";
+import { OpenAPI2Parameter, OpenAPI2Schema } from "../src/openapi2-document.js";
 import { diagnoseOpenApiFor, oapiForModel, openApiFor } from "./test-host.js";
 
 describe("typespec-autorest: primitives", () => {
